README.md: document LFS benefits
We get 98.85% space savings with git LFS. You download
2.5M instead of 218M if you use GIT_LFS_SKIP_SMUDGE=1,
so recommend it.

In the future CI's can leverage this as well, as they can
use it to just add their new test results and push. CIs
won't need to full repo. They'd have it if you have a local
mirror of the archive, of course. The main benefit for
space savings if for developers.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>